Original list (7/2016 DLF eResearch Network talk)
Here's the first list I put together, for a July 20, 2016 presentation introducing the digital humanities (what it is, why it can be good for you/your unit/your institution, how to get started in DH) to the DLF (Digital Library Federation) eResearch Network. The idea was that in addition to giving a definition of DH, to quickly step through a list of diverse examples of who does DH and what DH can be. I want people to come away inspired by the breadth of who is working in DH and what DH work is being done, and to feel enfranchised to include themselves and their work in DH should they choose. I was working within the limits of a 40-minute presentation and spoken form (vs. the text version now on my blog (http://literaturegeek.com/2016/07/21/dlf-digital-humanities-what-why-how).